+++ yarn-project/blob-sink/src/blobstore/memory_blob_store.test.ts
@@ -2,5 +2,5 @@ import { describeBlobStore } from './blob_store_test_suite.js';
 import { MemoryBlobStore } from './memory_blob_store.js';
 
 describe('MemoryBlobStore', () => {
-  describeBlobStore(() => new MemoryBlobStore());
+  describeBlobStore(() => Promise.resolve(new MemoryBlobStore()));
 });